refactor(newsletter): use shared digest layout, drop Trending tab and its server-side feed
